got the cycle going

+++ localbites_frontend/src/app/user/cart/cart.component.ts
@@ -3,6 +3,7 @@ import { Cart, Cart_Details } from '../../shared/models/Cart';
 import { ListingsService } from '../../listings.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Order, Order_Details } from '../../shared/models/Order';
+import { User } from '../../shared/models/User';
 
 @Component({
   selector: 'app-cart',
@@ -12,6 +13,8 @@ import { Order, Order_Details } from '../../shared/models/Order';
 export class CartComponent implements OnInit{
   cart!:Cart;
   cart_items: Cart_Details[] = [];
+  wallet!:number;
+  user!:User;
   user_id!:string;
   user_id_num!:number;
   orders: Order[] = [];
@@ -30,6 +33,11 @@ export class CartComponent implements OnInit{
       this.user_id = params.get('user_id')!;
       this.user_id_num = this.user_id !== null ? parseInt(this.user_id, 10) : 0;
       if (this.user_id) {
+        this.cartService.getUserSpecificAccount(this.user_id).subscribe(user => {
+          this.user = user;
+          this.wallet = user.wallet;
+        })
+
         this.cartService.getUserSpecificCartRoute(this.user_id).subscribe(cart => {
           // console.log("cart: " + cart)
           this.cart = cart;
@@ -89,40 +97,47 @@ export class CartComponent implements OnInit{
   }
 
   moveToPendingOrder(): void {
-    this.user_id_num = this.user_id !== null ? parseInt(this.user_id, 10) : 0; // Assuming default value of 0 if user_id is null
-    this.cartService.addOrderRoute(this.user_id_num, this.cart.total_price).subscribe({
-      next: (res) => {
-        console.log("order route worked: " + res.order_id)
-        this.orders.push(res)
-        this.cartService.getUserSpecificCartDetailsRoute(this.user_id).subscribe(cart_items => {
-          for (const cartItem of this.cart_items) {
-            // Call moveCartToPending for each item
-            this.cartService.moveCartToOrders(this.user_id_num, res.order_id, cartItem.name, cartItem.price, cartItem.item_id).subscribe(orders => {
-              console.log("moved to orders succesfully: " + res.order_id);
-                // Optionally, update the UI or do any additional processing
-                // this.cartService.deleteCartRoute(id).subscribe(() => {
-              this.order_details.push(orders)
+    if (this.user.wallet - this.cart.total_price > 0) {
 
-              },
-              error => {
-                console.error(`Failed to move cart details to orders: `, error);
-                // Optionally, handle the error or provide user feedback
-              }
-            );
-          }
-          this.cartService.deleteAllUserCartRoute(this.user_id).subscribe(() => {
-            this.cartService.updateCartByIdRoute(this.user_id, 0, 0).subscribe(() => {
-              this.cart_items = [];
-              this.cart.total_price = 0;
-              this.cart.quantity = 0;
+
+      this.user_id_num = this.user_id !== null ? parseInt(this.user_id, 10) : 0; // Assuming default value of 0 if user_id is null
+      this.cartService.addOrderRoute(this.user_id_num, this.cart.total_price).subscribe({
+        next: (res) => {
+          console.log("order route worked: " + res.order_id)
+          this.orders.push(res)
+          this.cartService.getUserSpecificCartDetailsRoute(this.user_id).subscribe(cart_items => {
+            for (const cartItem of this.cart_items) {
+              // Call moveCartToPending for each item
+              this.cartService.moveCartToOrders(this.user_id_num, res.order_id, cartItem.name, cartItem.price, cartItem.item_id).subscribe(orders => {
+                console.log("moved to orders succesfully: " + res.order_id);
+                  // Optionally, update the UI or do any additional processing
+                  // this.cartService.deleteCartRoute(id).subscribe(() => {
+                this.order_details.push(orders)
+
+                },
+                error => {
+                  console.error(`Failed to move cart details to orders: `, error);
+                  // Optionally, handle the error or provide user feedback
+                }
+              );
+            }
+            this.cartService.deleteAllUserCartRoute(this.user_id).subscribe(() => {
+              this.cartService.updateCartByIdRoute(this.user_id, 0, 0).subscribe(() => {
+                this.cart_items = [];
+                this.cart.total_price = 0;
+                this.cart.quantity = 0;
+              })
             })
-          })
-        });
-      },
-      error: (err) => {
-        console.error('Error adding restaurant: ', err);
-      }
-    });
+          });
+        },
+        error: (err) => {
+          console.error('Error adding restaurant: ', err);
+        }
+      });
+    }
+    else {
+      console.error('not enough funds')
+  } 
   }
 
 }
